Fix JRUBY-6952
Time#+ gives off-by-one error with fractional microseconds

I'm not sure why Time#+ used double throughout, where Time#- used
long, but changing + to all longs seems to fix this issue. I added
logging to see the values all along the way, and it's clear that
the long values were getting truncated inside double, leading to
this rounding error. With all longs, we should be losing no
precision (they're all integer values anyway) and should avoid
the rounding error.

+++ src/org/jruby/RubyTime.java
@@ -469,15 +469,15 @@ public class RubyTime extends RubyObject {
     }
 
     private IRubyObject opPlusNanos(long adjustNanos) {
-        double currentNanos = getTimeInMillis() * 1000000 + nsec;
+        long currentNanos = getTimeInMillis() * 1000000 + nsec;
 
-        double newNanos = currentNanos + adjustNanos;
-        double newMillisPart = newNanos / 1000000;
-        double newNanosPart = newNanos % 1000000;
+        long newNanos = currentNanos + adjustNanos;
+        long newMillisPart = newNanos / 1000000;
+        long newNanosPart = newNanos % 1000000;
 
         RubyTime newTime = new RubyTime(getRuntime(), getMetaClass());
         newTime.dt = new DateTime((long)newMillisPart).withZone(dt.getZone());
-        newTime.setNSec((long)newNanosPart);
+        newTime.setNSec(newNanosPart);
 
         return newTime;
     }
